everything workign except tours and annotations
