+++ src/style/global.css
@@ -4,7 +4,7 @@
     @apply bg-white
 }
 
-p > a {
+p > a:not(.a-button) {
   color: maroon;
   background-image: linear-gradient(0deg, rgba(0, 102, 136, 0.2) 10%, transparent 0);
 }
